Remove BusElement
