--- src/main/java/com/powsybl/openloadflow/network/LfBusElement.java
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.openloadflow.network;
-
-public interface LfBusElement {
-
-    LfBus getBus();
-
-    void setBus(LfBus bus);
-}
