+++ src/main/java/com/powsybl/openloadflow/network/LfGenerator.java
@@ -11,9 +11,13 @@ import java.util.OptionalDouble;
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public interface LfGenerator extends LfBusElement {
+public interface LfGenerator {
 
     String getId();
+
+    LfBus getBus();
+
+    void setBus(LfBus bus);
 
     boolean hasVoltageControl();
