--- src/main/java/com/powsybl/openloadflow/network/impl/AbstractLfBusElement.java
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.openloadflow.network.impl;
-
-import com.powsybl.openloadflow.network.LfBus;
-
-public abstract class AbstractLfBusElement {
-    protected LfBus bus = null;
-
-    public LfBus getBus() {
-        return bus;
-    }
-
-    public void setBus(LfBus bus) {
-        this.bus = bus;
-    }
-}
